+++ src/main/kotlin/net/ndrei/teslacorelib/render/selfrendering/IProvideVariantTransform.kt
@@ -1,0 +1,7 @@
+package net.ndrei.teslacorelib.render.selfrendering
+
+import net.minecraftforge.common.model.TRSRTransformation
+
+interface IProvideVariantTransform {
+    fun getTransform(variant: String): TRSRTransformation
+}